Fixed some errors in replying to messages.
Replying to a message would fail due to an undefined local variable
'file', which appears to be a typo for msg.file. cmd_mark was also
improperly called with more than one argument, this was also fixed.

I also received an error due to a nil depth in niceobj, so I put a guard
on it. This prevents the crashing symptom, but the cause of the nil
depth has yet to be found.